Image Processing modules ( mean & median filter) using MPI
The idea is that the application will read in an image, partition it
into a set of subimages, and distribute the subimages to N processors,
where each of processors will perform the image processing routine(s).
The number of processors will be determined at runtime. The worker
processors will return their portion to the master node, where subimages
are assembled back to a single large and processed image.